requirements.txt
pyyaml
watchdog
hnswlib; sys_platform != 'win32'
networkx
scikit-learn
numpy
pandas
rapidfuzz
python-dotenv
rank-bm25
python-louvain

sentencepiece
transformers>=4.43
accelerate>=0.33
sentence-transformers>=3.0
regex
unidecode

python-magic-bin; sys_platform == 'win32'
PyMuPDF
chardet
# pytesseract  # optional